Move KeyManager permissions to own component
